refactor HadesScheduler docker file structure
