+++ HadesScheduler/docker/scheduler.go
@@ -1,0 +1,153 @@
+package docker
+
+import (
+	"context"
+	"fmt"
+	"log/slog"
+	"strconv"
+	"time"
+
+	"github.com/docker/docker/api/types/container"
+	"github.com/docker/docker/client"
+	"github.com/ls1intum/hades/hadesScheduler/fluentd"
+	"github.com/ls1intum/hades/hadesScheduler/log"
+	"github.com/ls1intum/hades/shared/payload"
+	"github.com/ls1intum/hades/shared/utils"
+	"github.com/nats-io/nats.go"
+
+	slogfluentd "github.com/samber/slog-fluentd/v2"
+)
+
+type DockerEnvConfig struct {
+	DockerHost           string `env:"DOCKER_HOST" envDefault:"unix:///var/run/docker.sock"`
+	ContainerAutoremove  bool   `env:"DOCKER_CONTAINER_AUTOREMOVE" envDefault:"false"`
+	DockerScriptExecutor string `env:"DOCKER_SCRIPT_EXECUTOR" envDefault:"/bin/bash -c"`
+	CPU_limit            uint   `env:"DOCKER_CPU_LIMIT"`    // Number of CPUs - e.g. '6'
+	MEMORY_limit         string `env:"DOCKER_MEMORY_LIMIT"` // RAM usage in g or m  - e.g. '4g'
+}
+
+type DockerProps struct {
+	scriptExecutor       string
+	containerAutoremove  bool
+	cpu_limit            uint
+	memory_limit         string
+	volumeName           string
+	containerLogsOptions container.LogConfig
+}
+
+type Scheduler struct {
+	cli *client.Client
+	DockerProps
+	fluentd.FluentdOptions
+	publisher log.Publisher
+}
+
+func NewDockerScheduler() *Scheduler {
+	var dockerCfg DockerEnvConfig
+	utils.LoadConfig(&dockerCfg)
+	slog.Debug("Docker config", "config", dockerCfg)
+
+	var err error
+	// Create a new Docker client
+	cli, err := client.NewClientWithOpts(client.WithHost(dockerCfg.DockerHost), client.WithAPIVersionNegotiation())
+	if err != nil {
+		slog.Error("Failed to create Docker client", slog.Any("error", err))
+	}
+	return &Scheduler{
+		cli: cli,
+		DockerProps: DockerProps{
+			scriptExecutor:      dockerCfg.DockerScriptExecutor,
+			containerAutoremove: dockerCfg.ContainerAutoremove,
+			cpu_limit:           dockerCfg.CPU_limit,
+			memory_limit:        dockerCfg.MEMORY_limit,
+		},
+	}
+}
+
+func (d *Scheduler) SetFluentdLogging(addr string, max_retries uint) *Scheduler {
+	if addr != "" {
+		d.FluentdOptions = fluentd.FluentdOptions{
+			Addr:     addr,
+			MaxRetry: max_retries,
+		}
+	}
+	return d
+}
+
+func (d *Scheduler) SetNatsConnection(nc *nats.Conn) *Scheduler {
+	if nc != nil {
+		d.publisher = log.NewNATSPublisher(nc)
+	} else {
+		slog.Warn("NATS connection is nil, logs will not be published")
+	}
+	return d
+}
+
+func (d Scheduler) ScheduleJob(ctx context.Context, job payload.QueuePayload) error {
+	// Create a custom logger only for this job when fluentd is enabled
+	var job_logger *slog.Logger
+	var container_logs_options container.LogConfig
+	if d.FluentdOptions.Addr != "" {
+		fluentd_client, err := fluentd.GetFluentdClient(d.FluentdOptions)
+		if err != nil {
+			slog.Error("Failed to create fluentd client", slog.Any("error", err))
+			return err
+		}
+		job_logger = slog.New(slogfluentd.Option{
+			Level:  slog.LevelDebug,
+			Client: fluentd_client,
+		}.NewFluentdHandler()).With(slog.String("job_id", job.ID.String()))
+
+		// Configure the handler for the container logs
+		container_logs_options = container.LogConfig{
+			Type: "fluentd",
+			Config: map[string]string{
+				"fluentd-address":     d.FluentdOptions.Addr,
+				"fluentd-max-retries": strconv.FormatUint(uint64(d.FluentdOptions.MaxRetry), 10),
+				"labels":              "job_id",
+			},
+		}
+	} else {
+		job_logger = slog.Default().With(slog.String("job_id", job.ID.String()))
+		container_logs_options = container.LogConfig{}
+		job_logger.Warn("No fluentd address provided, using default logger")
+	}
+	// Create a unique volume name for this job
+	volumeName := fmt.Sprintf("shared-%s", job.ID.String())
+	// Create the shared volume
+	if err := createSharedVolume(ctx, d.cli, volumeName); err != nil {
+		job_logger.Error("Failed to create shared volume", slog.Any("error", err))
+		return err
+	}
+
+	// Add created volume to the job's docker config
+	jobDockerConfig := d.DockerProps
+	jobDockerConfig.volumeName = volumeName
+	jobDockerConfig.containerLogsOptions = container_logs_options
+	docker_job := DockerJob{
+		cli:          d.cli,
+		logger:       job_logger,
+		DockerProps:  jobDockerConfig,
+		QueuePayload: job,
+		publisher:    d.publisher,
+	}
+	err := docker_job.execute(ctx)
+	if err != nil {
+		job_logger.Error("Failed to execute job", slog.Any("error", err))
+		return err
+	}
+
+	job_logger.Debug("Job executed successfully", slog.Any("job_id", job.ID))
+
+	// Delete the shared volume after the job is done
+	defer func() {
+		time.Sleep(1 * time.Second)
+		if err := deleteSharedVolume(ctx, d.cli, volumeName); err != nil {
+			job_logger.Error("Failed to delete shared volume", slog.Any("error", err))
+		}
+
+		job_logger.Info("Volume deleted", slog.Any("volume", volumeName))
+	}()
+
+	return nil
+}
